[WIP] Notification API
Proof of concept of notification API, which currently aims to
provide possibility of sending signals to the containers. The
concept of notifications isn't signal-centered though.

Also, the "kill" word which was used in context of graceful
deletion of containers, was changed to "stop", in order to
avoid confusion with the "kill" word used in context of
custom signals.

Ref #24957